requirements.txt
# --- 1. 기본 수학/데이터 (Numpy 2.0 호환성 문제 방지) ---
numpy<2.0.0,>=1.24.1
pandas
scipy
matplotlib
seaborn
tqdm

# --- 2. 설정 및 유틸리티 ---
yacs
easydict
termcolor
prettytable
typing_extensions
Pillow

# --- 3. AI/Deep Learning 보조 ---
# (torch는 1단계에서 따로 설치했으므로 생략하거나 주석 처리)
timm
thop
tensorboard
graphviz

# --- 4. NADER 핵심 (NAS & LLM) ---
nas-bench-201
openai>=1.33.0
langchain>=0.2.1
langchain-community
langchain-core
langchain-classic
langchain-openai
langchain-chroma
chromadb
tiktoken

# --- 5. OpenMMLab (설치 에러 시 제외 가능) ---
mmengine
mmcv-lite>=2.0.0rc4